+++ domain/config.go
@@ -161,6 +161,10 @@ var DefaultConfig = Config{
 		ServerConnectionTimeoutSeconds: 10,
 		Plugins: []Plugin{
 			&OrderBookPluginConfig{
+				Enabled: true,
+				Name:    orderbookplugindomain.OrderbookOrdersCachePlugin,
+			},
+			&OrderBookPluginConfig{
 				Enabled: false,
 				Name:    orderbookplugindomain.OrderbookFillbotPlugin,
 			},
@@ -385,6 +389,8 @@ func validateDynamicMinLiquidityCapDesc(values []DynamicMinLiquidityCapFilterEnt
 // PluginFactory creates a Plugin instance based on the provided name.
 func PluginFactory(name string) Plugin {
 	switch name {
+	case orderbookplugindomain.OrderbookOrdersCachePlugin:
+		return &OrderBookPluginConfig{}
 	case orderbookplugindomain.OrderbookFillbotPlugin:
 		return &OrderBookPluginConfig{}
 	case orderbookplugindomain.OrderbookClaimbotPlugin: