+++ domain/orderbook/orderbook_repository.go
@@ -15,4 +15,10 @@ type OrderBookRepository interface {
 	// GetTickByID returns a specific orderbook tick for a given orderbook pool id.
 	// Returns false if the tick is not found.
 	GetTickByID(poolID uint64, tickID int64) (OrderbookTick, bool)
+
+	// StoreOrders stores the orders for a given orderbook pool id.
+	StoreOrders(poolID uint64, orders []Order) error
+
+	// GetOrders returns the orders for a given orderbook pool id.
+	GetOrders(poolID uint64, ownerAddress string) ([]Order, bool)
 }