+++ CHANGELOG.md
@@ -35,10 +35,23 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 # Changelog
 
-## Unreleased
+## v28.3.3
+
+- #622 - feat: orderbook orders cache plugin
+
+## v28.3.2
+
+- #638 - BE-714 | Fix candidate_routes_compute_duration gauge call
+
+## v28.3.1
+
+- #637 - BE-714 | Optimize route finding algorithm
+
+## v28.3.0
 
 - Remove dependency on keyring, wire in CosmosSigner and grab private key from SQS_PRIVATE_KEY environment variable.
-- #632 - E-697 | Increase max routes for pricing worker
+- #632 - BE-697 | Increase max routes for pricing worker
+- #633 - BE-707 | Skip processing quote requests when system is not healthy 
 
 ## v28.2.0